requirements.txt
optuna
scikit-learn
numpy
pandas
matplotlib
xgboost
lightgbm
umap-learn
optuna
